Move tests to their own submodule
